regen

+++ typespec-tests/src/main/java/com/type/scalar/Decimal128TypeAsyncClient.java
@@ -17,6 +17,7 @@ import com.azure.core.http.rest.Response;
 import com.azure.core.util.BinaryData;
 import com.azure.core.util.FluxUtil;
 import com.type.scalar.implementation.Decimal128TypesImpl;
+import java.math.BigDecimal;
 import reactor.core.publisher.Mono;
 
 /**
@@ -43,7 +44,7 @@ public final class Decimal128TypeAsyncClient {
      * <strong>Response Body Schema</strong>
      * </p>
      * <pre>{@code
-     * double
+     * BigDecimal
      * }</pre>
      * 
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
@@ -65,7 +66,7 @@ public final class Decimal128TypeAsyncClient {
      * <strong>Request Body Schema</strong>
      * </p>
      * <pre>{@code
-     * double
+     * BigDecimal
      * }</pre>
      * 
      * @param body A 128-bit decimal number.
@@ -95,7 +96,7 @@ public final class Decimal128TypeAsyncClient {
      */
     @Generated
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Response<Void>> requestParameterWithResponse(double value, RequestOptions requestOptions) {
+    public Mono<Response<Void>> requestParameterWithResponse(BigDecimal value, RequestOptions requestOptions) {
         return this.serviceClient.requestParameterWithResponseAsync(value, requestOptions);
     }
 
@@ -111,11 +112,11 @@ public final class Decimal128TypeAsyncClient {
      */
     @Generated
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Double> responseBody() {
+    public Mono<BigDecimal> responseBody() {
         // Generated convenience method for responseBodyWithResponse
         RequestOptions requestOptions = new RequestOptions();
         return responseBodyWithResponse(requestOptions).flatMap(FluxUtil::toMono)
-            .map(protocolMethodData -> protocolMethodData.toObject(Double.class));
+            .map(protocolMethodData -> protocolMethodData.toObject(BigDecimal.class));
     }
 
     /**
@@ -132,7 +133,7 @@ public final class Decimal128TypeAsyncClient {
      */
     @Generated
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Void> requestBody(double body) {
+    public Mono<Void> requestBody(BigDecimal body) {
         // Generated convenience method for requestBodyWithResponse
         RequestOptions requestOptions = new RequestOptions();
         return requestBodyWithResponse(BinaryData.fromObject(body), requestOptions).flatMap(FluxUtil::toMono);
@@ -152,7 +153,7 @@ public final class Decimal128TypeAsyncClient {
      */
     @Generated
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Void> requestParameter(double value) {
+    public Mono<Void> requestParameter(BigDecimal value) {
         // Generated convenience method for requestParameterWithResponse
         RequestOptions requestOptions = new RequestOptions();
         return requestParameterWithResponse(value, requestOptions).flatMap(FluxUtil::toMono);
